+++ src/elements/components/link.module.scss
@@ -5,7 +5,7 @@
         text-decoration: underline;
     }
 
-    &.external *:last-child {
+    .externalIcon {
         margin-bottom: -0.125em;
         margin-left: 0.1em;
         margin-right: 0.1em;
